replies clear

+++ forum-api/src/Applications/use_case/replies/DeleteReplyUseCase.js
@@ -1,12 +1,19 @@
 const DeleteReply = require('../../../Domains/replies/entites/DeleteReply');
 
 class DeleteReplyUseCase {
-  constructor({ replyRepository }) {
+  constructor({ replyRepository, threadRepository, commentRepository }) {
     this._replyRepository = replyRepository;
+    this._threadRepository = threadRepository;
+    this._commentRepository = commentRepository;
   }
 
   async execute(useCasePayload) {
-    const deleteReply = new DeleteReply(useCasePayload);
+    const { threadId, commentId, replyId, owner } = useCasePayload;
+    const deleteReply = new DeleteReply({ id: replyId, owner });
+    await this._threadRepository.verifyThreadAvailability(threadId);
+    await this._commentRepository.verifyCommentAvailability(commentId);
+    await this._replyRepository.verifyReplyOwner(replyId, owner);
+
     const deletedReply = await this._replyRepository.deleteReply(deleteReply);
     return deletedReply;
   }
